Alpha channel support, don't use inventory cubes
Add alpha channel support for latest git clients/servers
(should still work fine on old clients/servers too).

Requires that the client use shaders to see the effect.

Also, removes the calls that create rendered inventory cubes - those
calls are obsolete and should not be used in production, as they prevent
the engine from rendering an actual cube when an item is dropped.